Rename back from subtensor_module.py to subtensor.py
