+++ bittensor/__init__.py
@@ -286,7 +286,7 @@ from .chain_data import (
     ProposalCallData,
     ProposalVoteData,
 )
-from .subtensor_module import subtensor as subtensor
+from .subtensor import subtensor as subtensor
 from .cli import cli as cli, COMMANDS as ALL_COMMANDS
 from .btlogging import logging
 from .metagraph import metagraph as metagraph
