+++ tests/unit_tests/extrinsics/test_registration.py
@@ -1,6 +1,6 @@
 import pytest
 from unittest.mock import MagicMock, patch
-from bittensor.subtensor_module import subtensor as Subtensor
+from bittensor.subtensor import subtensor as Subtensor
 from bittensor.wallet import wallet as Wallet
 from bittensor.utils.registration import POWSolution
 from bittensor.extrinsics.registration import (
